Change int to size_t
Uniformity Andi think it haS reasonable bounds if not just an alias

+++ getlapsbof.cpp
@@ -60,7 +60,7 @@ extern "C" {
         return len;
     }
 
-    int strcmp(const char *str1, const char *str2) {
+    size_t strcmp(const char *str1, const char *str2) {
         while (*str1 && (*str1 == *str2)) {
             str1++;
             str2++;
